requirements.txt
# --- Core ---
python-dotenv==1.0.1
requests==2.32.3

# --- Supabase & DB ---
supabase==2.4.5
psycopg2-binary==2.9.9

# --- Telegram Bot ---
python-telegram-bot==20.3

# --- API Backend ---
fastapi==0.115.0
uvicorn==0.30.6

# --- Machine Learning ---
scikit-learn==1.5.2
numpy==2.1.1
joblib==1.4.2

# --- Extra Utils ---
pydantic==2.9.2
openai==1.45.0